+++ src/main/java/mrfast/sbt/features/stats/number/ManaNumber.kt
@@ -8,28 +8,28 @@ import mrfast.sbt.utils.LocationUtils
 import mrfast.sbt.utils.Utils
 import mrfast.sbt.utils.Utils.formatNumber
 
-object OverflowManaDisplay {
+object ManaNumber {
     init {
-        OverflowManaDisplayGui()
+        ManaNumberGui()
     }
 
-    class OverflowManaDisplayGui : GuiManager.Element() {
+    class ManaNumberGui : GuiManager.Element() {
         init {
             this.relativeX = 0.371875
             this.relativeY = 0.842593
-            this.elementName = "Overflow Mana Display"
+            this.elementName = "Mana Number"
             this.addToList()
             this.height = Utils.mc.fontRendererObj.FONT_HEIGHT
         }
 
         override fun draw() {
-            val display = "§3${PlayerStats.overflowMana.formatNumber()}ʬ"
-            GuiUtils.drawText(display, 0f, 0f, GuiUtils.TextStyle.BLACK_OUTLINE)
-            this.width = Utils.mc.fontRendererObj.getStringWidth(display)+1
+            val number = "§9${PlayerStats.mana.formatNumber()}/${PlayerStats.maxMana.formatNumber()}"
+            GuiUtils.drawText(number, 0f, 0f, GuiUtils.TextStyle.BLACK_OUTLINE)
+            this.width = Utils.mc.fontRendererObj.getStringWidth(number) + 1
         }
 
         override fun isActive(): Boolean {
-            return GeneralConfig.overflowManaDisplay && LocationUtils.inSkyblock
+            return GeneralConfig.manaNumber && LocationUtils.inSkyblock
         }
 
         override fun isVisible(): Boolean {
