+++ src/main/java/mrfast/sbt/features/stats/number/OverflowManaNumber.kt
@@ -1,36 +1,35 @@
 package mrfast.sbt.features.stats.number
 
+import mrfast.sbt.apis.PlayerStats
 import mrfast.sbt.config.categories.GeneralConfig
 import mrfast.sbt.config.GuiManager
 import mrfast.sbt.utils.GuiUtils
 import mrfast.sbt.utils.LocationUtils
 import mrfast.sbt.utils.Utils
 import mrfast.sbt.utils.Utils.formatNumber
-import net.minecraft.client.Minecraft
 
-object SpeedDisplay {
+object OverflowManaNumber {
     init {
-        SpeedDisplayGui()
+        OverflowManaNumberGui()
     }
 
-    class SpeedDisplayGui : GuiManager.Element() {
+    class OverflowManaNumberGui : GuiManager.Element() {
         init {
             this.relativeX = 0.371875
             this.relativeY = 0.842593
-            this.elementName = "Speed Display"
+            this.elementName = "Overflow Mana Number"
             this.addToList()
             this.height = Utils.mc.fontRendererObj.FONT_HEIGHT
         }
 
         override fun draw() {
-            val speed = ((Minecraft.getMinecraft().thePlayer?.capabilities?.walkSpeed ?: 0f) * 1000).toFloat();
-            val display = "§r${speed.formatNumber()}%"
-            GuiUtils.drawText(display, 0f, 0f, GuiUtils.TextStyle.BLACK_OUTLINE)
-            this.width = Utils.mc.fontRendererObj.getStringWidth(display)+1
+            val number = "§3${PlayerStats.overflowMana.formatNumber()}ʬ"
+            GuiUtils.drawText(number, 0f, 0f, GuiUtils.TextStyle.BLACK_OUTLINE)
+            this.width = Utils.mc.fontRendererObj.getStringWidth(number) + 1
         }
 
         override fun isActive(): Boolean {
-            return GeneralConfig.speedDisplay && LocationUtils.inSkyblock
+            return GeneralConfig.overflowManaNumber && LocationUtils.inSkyblock
         }
 
         override fun isVisible(): Boolean {
