+++ src/main/java/mrfast/sbt/features/stats/number/SpeedNumber.kt
@@ -1,35 +1,36 @@
 package mrfast.sbt.features.stats.number
 
-import mrfast.sbt.apis.PlayerStats
 import mrfast.sbt.config.categories.GeneralConfig
 import mrfast.sbt.config.GuiManager
 import mrfast.sbt.utils.GuiUtils
 import mrfast.sbt.utils.LocationUtils
 import mrfast.sbt.utils.Utils
 import mrfast.sbt.utils.Utils.formatNumber
+import net.minecraft.client.Minecraft
 
-object ManaNumber {
+object SpeedNumber {
     init {
-        ManaNumberGui()
+        SpeedNumberGui()
     }
 
-    class ManaNumberGui : GuiManager.Element() {
+    class SpeedNumberGui : GuiManager.Element() {
         init {
             this.relativeX = 0.371875
             this.relativeY = 0.842593
-            this.elementName = "Mana Number"
+            this.elementName = "Speed Number"
             this.addToList()
             this.height = Utils.mc.fontRendererObj.FONT_HEIGHT
         }
 
         override fun draw() {
-            val number = "§9${PlayerStats.mana.formatNumber()}/${PlayerStats.maxMana.formatNumber()}"
+            val speed = ((Minecraft.getMinecraft().thePlayer?.capabilities?.walkSpeed ?: 0f) * 1000).toFloat();
+            val number = "§r${speed.formatNumber()}%"
             GuiUtils.drawText(number, 0f, 0f, GuiUtils.TextStyle.BLACK_OUTLINE)
-            this.width = Utils.mc.fontRendererObj.getStringWidth(number)+1
+            this.width = Utils.mc.fontRendererObj.getStringWidth(number) + 1
         }
 
         override fun isActive(): Boolean {
-            return GeneralConfig.manaDisplay && LocationUtils.inSkyblock
+            return GeneralConfig.speedNumber && LocationUtils.inSkyblock
         }
 
         override fun isVisible(): Boolean {
